feat: criando estrutura de paginação

+++ src/main/java/org/smartbar/aps_4semestre/pagination/PaginationDTO.java
@@ -10,6 +10,6 @@ public record PaginationDTO <E> (
 
         int size,
 
-        long totalPages
+        int totalPages
 
 ) {}
